refactored providers, we now use infura by default on node, and etherscan by default on the web
