--- libs/ethers-provider/src/development/index.ts
@@ -1,1 +1,0 @@
-export { Erc20Provider } from './provider';
