--- libs/ethers-provider/src/devlive/index.ts
@@ -1,2 +1,0 @@
-export { Erc20Provider } from './provider';
-export { deployProvider } from './deploy';
