+++ libs/ethers-provider/src/node/mocked.ts
@@ -1,0 +1,4 @@
+export const getProvider = () => {
+
+  throw new Error(`TODO: Implement mocked node provider`);
+}
