--- libs/ethers-provider/src/prod/index.ts
@@ -1,3 +1,0 @@
-export { Erc20Provider } from './provider';
-export type { ERC20Response } from './erc20response';
-export { deployProvider } from './deploy';
