+++ libs/ethers-provider/src/web/local/index.ts
@@ -2,7 +2,7 @@ import { BlockTag, Filter, JsonRpcProvider, Log } from '@ethersproject/providers
 import { hexZeroPad, hexStripZeros } from "@ethersproject/bytes";
 import { id } from "@ethersproject/hash";
 import { BigNumber } from "@ethersproject/bignumber";
-import { ERC20Response } from '../prod/erc20response';
+import { ERC20Response } from '../erc20response';
 
 export class Erc20Provider extends JsonRpcProvider {
 
@@ -53,3 +53,5 @@ export class Erc20Provider extends JsonRpcProvider {
     return result;
   }
 }
+
+export const getProvider = () => new Erc20Provider();
